refactor(habitat): plan source check files by coverage
