+++ tools/habitat-harness/src/domains/rule-registry/index.ts
@@ -1,4 +1,5 @@
 export * from "./facts.js";
 export { ruleGraphFacts } from "./graph.js";
 export * from "./load.js";
+export * from "./path-coverage.js";
 export * from "./schema.js";
