+++ tools/habitat-harness/src/domains/rule-registry/schema.ts
@@ -261,6 +261,7 @@ export const RulePatternFactsSchema = Type.Pick(PatternCheckRuleRegistryRecordV1
   "lane",
   "message",
   "patternName",
+  "pathCoverage",
   "scanRoots",
 ]);
