+++ tools/habitat-harness/src/domains/workspace-graph-integration/routing.ts
@@ -1,6 +1,7 @@
 import type { WorkspaceProject } from "../../providers/nx/schema.js";
 import { activeRuleRoutingFacts } from "../rule-registry/active-facts.js";
 import type { RuleRoutingFacts } from "../rule-registry/index.js";
+import { pathCoveragePatternMatches } from "../rule-registry/path-coverage.js";
 import type { RuleCoverageKind, RuleRouting } from "./schema.js";
 
 export function rulesForPath(pathInRepo: string, owner?: WorkspaceProject): RuleRouting[] {
@@ -59,45 +60,4 @@ function routingFact(
     coverageKind,
     reason,
   };
-}
-
-function pathCoveragePatternMatches(pattern: string, pathInRepo: string): boolean {
-  const normalized = pattern.replaceAll("\\", "/");
-  if (!normalized.includes("*")) {
-    return pathInRepo === normalized || pathInRepo.startsWith(`${normalized}/`);
-  }
-  return globToRegExp(normalized).test(pathInRepo);
-}
-
-function globToRegExp(pattern: string): RegExp {
-  let source = "^";
-  for (let index = 0; index < pattern.length; index += 1) {
-    const char = pattern[index];
-    const next = pattern[index + 1];
-    if (char === "*" && next === "*") {
-      if (pattern[index + 2] === "/") {
-        source += "(?:.*/)?";
-        index += 2;
-        continue;
-      }
-      source += ".*";
-      index += 1;
-      continue;
-    }
-    if (char === "*") {
-      source += "[^/]*";
-      continue;
-    }
-    if (char === "?") {
-      source += "[^/]";
-      continue;
-    }
-    source += escapeRegExp(char);
-  }
-  source += "$";
-  return new RegExp(source);
-}
-
-function escapeRegExp(char: string) {
-  return /[\\^$+?.()|[\]{}]/.test(char) ? `\\${char}` : char;
 }
